Updates LindbladOp -> ExpErrogenOp.
Replaces usage of the old LindbladOp (and LindbladDenseOp) with
ExpErrorgenOp (and ExpErrorgenDenseOp), sometimes within a
ComposedOp, since LindbladOp could contain a "unitary post-factor".
The newer paradigm, whereby the LindbladErrorgen is the only object
that specifically has a Lindblad-like set of errors, caused us to
move (in this commit) the from_operation_matrix(...) classmethod
from what was LindbladOp to LindbladErrorgenOp.  Now it takes more
steps to construct a Lindblad-parameterized operation, but the
code seems clearer and is less redundant.