Switches mapforwardsim.py to use new self.calclib calls instead of replib.
This update allows the mapforwardsim to be used, at least for the
single existing evotype ('densitymx').  This commit also adds Evotype
casting, so that a model holds an Evotype object instead of just a
string, and fixes some bugs so that a simple 1Q GST test with the
map fwdsim works now and gives the same answer as the matrix fwdsim.