Adds default evotype plumbing and gets densitymx_slow evotype working.
Adds a `default_evotype` class variable to Evotype and uses this
value whenever "default" is given as an evotype - ideally anywhere
within pyGSTi but there are probably places this commit misses.
This allows the user to easily change what evotype they're working
with.  This commit updates many of the model creation functions
and those functions downstream, and updates docstrings to be more
generic and mention the default behavior.  The initial value of
`default_evotype` is either "densitymx" or "densitymx_slow",
depending on whether the former could be successfully imported.

Previously `parameterization` arguments could be used to specify
a parameterization type *and* and evotype, e.g. "H+S terms".  With
this commit this is no longer possible (the evotype really needs
its own argument) and evotype arguments have been added to cloud-
and local-noise model creation functions so that the evotype can
be passed in separately.

Finally, this commit updates the densitymx_slow evotype files
to match the densitymx ones (fixing the same bugs we fixed there)
so that now the densitymx_slow evotype at least works in simple
cases.