Adds 'state_space' attribute to all model members and representations.
Replaces the old and inconsistent dimension (dim) and nqubits with
a StateSpace object, which can hold the state space multiple more
or less efficient/flexible ways.  This commit touches many files and
bugs will undoubtably will need to be fixed later, but the upshot is
that model member objects and their representations now all contain
a state_space attribute that describes abstractly the space that the
object acts upon.  The StateSpace class itself still needs to be
implemented (statespace.py contains just stubs at this point), but
the plan is that a StateSpace object contains a superoperator dimension
(dim), a unitary operator dimension (udim) and, when appropriate, a
number of qubits (num_qubits), and operations interact with whatever
attribute of the state space they need, e.g. an rep object for the
'statevec' evotype might use the 'udim' whereas a 'densitymx' rep
might use 'dim'.

This commit also adds the StaticPureState class to mirror a
StaticState but representing a pure state that has no parameters.