Updates __init__ signatue of dense/dense-unitary op reps and renames "pure" -> "dense unitary"
Adds a 'mx' (dense matrix) argument to dense operator representation classes,
anticipating that this will be useful when adding the pathintegral evotype soon.
`None` can be passes as this matrix so the caller doesn't need to specify an
initial matrix if one isn't available.

This commit also finishes implementations of the pure/unitary versions
of operations, and renames "pure" to "dense unitary" in names regarding
operations, as this seems more appropriate.  In particlar, there now
exists a DenseUnitaryOperator base class (similar to DenseOperator but
for unitary ops) and a new StaticUnitaryOp class.