Updates model construction to use new evotypes & many downstream udpates.
One particular downstream update was the addition of 'basis'
arguments to several/many of the state- and op-rep classes.
A basis is needed whenever a representaion doesn't have enough
information to render itself as a dense superoperator (or bra or
ket), since an evotype (the densitymx type, for example) would
need this information in order to create or act on states.  In
cases when the state-rep doesn't hold a dense super-ket representation,
this basis isn't really needed (it would only be needed for a to_dense()
call).  The fact that a basis must be given may currently present a
performance bottleneck, as computing a default basis means getting the
dimension of a state space.  In the future, we should be able to construct
lightweight basis objects that are constructed ("cast") from a basis name
and a state space (e.g. just a number of qubits).

This commit makes a first pass at updating model construction, but
this code needs an udpate anyway so I didn't go crazy trying to
cover all the cases exactly.  We need to decide on a nice (common)
API for constructing cloud and local noise models, and in particular
what values of "parameterization" are allowed.  Use of "clifford" as
a parameterization seems dubious, but I think this functionality is
needed for some of the RB circuit compilation code, so we'll need to
check on this.