Updates Basis objects to lazily evaluation their dimension, size, and element-shape.
Since, after the evotype update and refactor a lot more objects carry around
a Basis object (e.g. so that a state vector can be transformed into a superop)
but may never actually need the basis for anything.  These basis objects may
need to represent bases for many (100s) of qubits, in which case computing
the basis dimension, size, etc., would result in integer overflows.  This
commit updates the Basis object so that its dimension, size, and other
would-be-overflowed quantities are computed lazily -- and so never need
to be computed in the 100s of qubits cases when we'll never ask for dense
representations of operators.

More work needs to be done in the Basis.cast method, which still reverts
to computing dims more often than it should.