Adds 'on_space' argument to to_dense methods.
This addition adds clarity within the evotype code, where previously
to_dense would behave differently depending on the evotype, sometimes
where this would be undesired.  The new on_space argument can be set
to 'Hilbert' (for Hilbert-space matrices/vectors), 'HilbertSchmidt'
(likewise) and 'minimal'.  The 'minimal' option uses the evotype's
natural dense representation, which will be a Hilbert-space matrix/vector
if the evotype is restricted to pure states and a Hilbert-Schmidt-space
matrix/vector otherwise.  I expect that some tweaking of the various
on_space arguments will be needed to get pure-vector simulations up
and going, as, for instance, the reporting code now demands 'HilbertSchmidt'
matrices/vectors and these may not always be available.