Renames *DenseOp classes to *ArbitraryOp and CliffordOp to StaticCliffordOp
In an attempt to have a consistent naming of the operation classes,
and to reflect the design that the operation/spam objects are
evotype-agnostic, this commit renames:

CliffordOp -> StaticCliffordOp
FullDenseOp -> FullArbitraryOp
TPDenseOp -> FullTPOp
StaticDenseOp -> StaticArbitraryOp
LinearlyParamDenseOp -> LinearlyParamArbitraryOp

An operation class defines 1) the parameterization of the op,
and 2) the class of possible operations that the object can represent.
For example, the "FullTPOp" object can represent any trace-preserving map
and parameterizes this map fully (meaning to the maximum extent it can).