Removes legacyio.py imports; adds a forgotten function to slowopcalc.py
Things have changed so much that we don't plan on updating legacyio
so that current pickles are compatible with those of the last version.
This commit removes the legacyio.py functions from pygsti.io.

This commit also adds the float_prod function, which existed in
fastopcalc.pyx, to slowopcalc.py.