Manually merged updates from outdated feature-noiseop-spamvec branch.
The recent large refactor duplicated some of the work in an existing
(feature-noiseop-spamvec) branch and it would have been difficult for
git to merge it correctly.  This commit is the result of taking all
the new commits on that branch and applying/adjusting them on top
of the post-refactored code.

The outdated branch created composed state and POVM objects,
integrated them into local-noise model construction, and made some
additional updates for running CHP (added sparse outcomes for
 bulk_probs in WeakForwardSim).  Since the large refactor also created
composed state and POVM objects these needed to be merged carefully;
the CHP enhancements could just be applied directly (easy to merge).
The unit tests from the old branch have also been included & updated
in this commit, and all of them pass now, like they did before.