Converts relative -> absolute imports and auto-indents many test & script files.
This commit marks the beginning of a second refactor, where we plan to
distribute the contents of the current 'construction' sub-package into
other sub-packages.  This commit does some preparatory work, by changing
most of the relative import statements within pygsti into absolute imports
(not necessary but more PEP8-compliant and arguably easier to read & maintain),
and includes the result of auto-indenting most if not all of the repo files,
which affected the tests and script files in particular (the main package
files were already indented nicely).