Adds new create_explicit_model function & replaces Models w/ProcessorSpecs in exp designs.
Moves the old explicit-model creation function to create_explicit_model_from_expressions,
and adds a new method that uses the new model-creation API.  There's still work to be
done regarding how to process local-noise items in the `modelnoise` argument (e.g.
'Gx' noise rather than ('Gx', 0) noise), but the essential structure is there now.

With this function, ProcessorSpecs can be converted into explict models, and when the
parameterization isn't essential -- as in the case of experiment designs -- it allows
us to replace an explict model with a mere ProcessorSpec.  This commit does this for
the `target_model` contained in GST experiment designs, so that now these designs
only hold a processor spec and a new `create_target_model` method uses that processor
spec to create a static explicit model.

As a side effect of implementing the new create_explicit_model function, this
commit updates what I think was an incorrect/incomplete implementation of
EmbeddedOp objects, particularly for the state-vector evotypes.  The statevec
and statevec_slow evotypes' implementation of embedded reps was broken, using
the superoperator dimension rather than the unitary dimension in places.  Also
the EmbeddedOp methods for iterating over the embedded elements (used for
to_dense and to_sparse calls) needed to be fixed/finished to use their
'on_space' arguments to properly perform the embedding in either the unitary
or superop space as desired.