Fixes and updates related to unit tests (initial pass).
Begins getting the unit tests to run by fixing bugs & mismatches
in the unit tests.  Mostly minor bug fixes and updates to the unit
tests, but some added features for making models behave like they
used to and for utilizing processorspecs in experiment designs
(this needs more work).  Many unit tests pass now, but not all of
them -- a WIP.