Updates the way implied idle gates are constructed in model creation functions.
By "implied" idle gates we mean the idles in a circuit layer that correspond
to empty/"blank" qubit lines.  Previously, implicit models would construct
and store special idle operations for use within their layer rules, and
these idle ops would be specified independently of the processor spec
(via arguments of the model creation fns).  This was awkward since
idle gates can also be specified as part of the processor spec.

Now, to attempt to remedy this issue, model objects don't have
hard-coded special idle ops but rather hold the name of idle
gates that are used to implement implied idles via the layer
rules.  Model construction/__init__ routines find, if any exist, idle
operations as specified in the processor spec and "link" these
names with the layer rules.  So now the processor spec *should*
specify any idle gates - e.g. a 1Q idle op and/or a global idle op.

A LocalNoiseModel looks for a global or 1Q idle op and uses this
to construct a global idle op that is used in layers.  A
CloudNoiseModel is similar, but currently only looks for a global
idle op.  An ExplicitOpModel doesn't have any layer rules, but
now has the special behavior that if there is a global idle op
whose name is paraenthesized (an "implicit gate name") then this
idle op is given the label () -- an empty tuple -- in the constructed
model.

This commit also adds, but doesn't utilize yet, functionality whereby
a ProcessorSpec can be given an integer # of qubits instead of a
unitary to indicate an idle operation that should only be built
if it has noise on it -- if there's no noise then no op should be
created.  This is particularly intended for creating global idle
operations, where we don't want/need to ever construct a n-qubit
unitary (e.g. for n > 50).

Finally, this commit adds convenience functionality to ProcessorSpec
so that if no unitary can be found for a gate name, and that gate
name contains the string 'idle', then the gate is assumed to be a
n-qubit idle gate where n is determined by the given availability
or defaults to n=total_number_of_qubits when no availability is given.

This improved idle processing should be more transparent to users
and leaves open more advanced ways of handling implied idles within
the implicit models.  It also rightly draws the idle gates from the
processor spec so that a processor spec can construct, e.g., an
explicit model *including* its idle gate.