Relaxes the truncation tolerance in ExpErrorgenOp.spam_transform_inplace from 1e-6 to 3e-6.
A unit test was failing (only on Python 3.6) because during gauge optimization
we were getting a negative eigenvalue of -1.04e-6 and the truncation level
was set to 1e-6.  Stepping this up to 3e-6, which should still be fine, and
gives a little more leeway for numerical errors when spam-transforming CPTP-
constrained, i.e, exp(Lindblad_errorgen) maps (ExpErrorgenOp objects).