Adds 'actionable_staterep' method to densitymx evotype StateRep objects.
This was forgotten, and isn't needed when the mapforwardsim_calc_densitymx.pyx
Cython module is built, but it seems not to be at least for Python 3.6.  A
brief assessment didn't reveal why this is, or if *any* of the GitHub builds
are compiling the Cython correctly, but for now we just patch the bug so that
things should work whether or not the Cython module is compiled.