Fixes ability to use densitymx_slow with cmputational basis states.
Updates a hardcoded test for "densitymx" that resulted in state.to_dense()
incorrectly returning a *complex* array in the densitymx_slow evotype and
thereby causing an assertion error in the Rep class, ultimately making GST
inoperable.  This commit fixes this issue, and also removes several assertions
for hard-coded evotypes that aren't appropriate anymore.