Removes portions of system tests that test plotly polar plots, which don't work now.
Due to a Plotly API change that we don't want to deal with yet, the polar eigenvalue
plots (which aren't currently used in reports) don't work anymore, failing with a
"invalid argument" from plotly.  This commit comments out the places where we tested
the polar plots in test_packages/reportb/testWorkspace.py.