Adds docstrings to many recently added modules.  Minor function signature updates.
Adds or updates docstrings to model construction, model noise,
stencil labels, and other modules to be consistent with recent
updates to the code.

The only non-docstring update was to change the
OpModelNoise.create_[errorgen|errormap]_stencil methods so that
they only take a number (not an actual set of) target labels.  Previously
a set of target labels was given that would be hard-coded into the "stencil",
which worked fine and was used for the "stencils" for specific labels like
('Gx',0).  It seems more intuitive, however, to just give the stencil the
*number* of target labels and have it default to using special references
to these, i.e. ('@0', '@1', ...) which then get replaced by the actual target
labels when the stencil is applied.  So this doesn't (or shouldn't) change
the way anything runs and actually adds computational overhead to make the
code clearer.