Initial implementation of parameter bounds in the "customlm" optimizer.
Adds a commented-out version of enforcing parameter bounds that reduces
the step size (length of dx vector) until all the components of the resulting
"new" x are valid, in case this seems more desirable later.

Adds what I expect to be the better (and simpler) method of enforcing bounds,
which just clips out-of-range components to the allowed range.  This is the
currently used method.

Both methods use a Model's new .parameter_bounds property, which currently
can just be set directly by the user but may have functionality similar to
paraemters in the future (where bounds are inherited from bounds in a
model member object).