Adds barrier or "compilable-layer" metadata to circuits.
Circuit objects are given a `compilable_layer_indices` attribute that
holds indices for which "compilation", i.e., combination with other
layers and the absence of a barrier after the layer.  This information
is also stored within circuit string representations via the '~' and '|'
(following a layer to mark it as "compilable" or being followed by a
barrier respectively).  The compilable layer indices should be updated
when layers are inserted or deleted, but this has not been tested
extensively and probably breaks with CircuitLabels (sub-circuits).  Some
basic unit tests are added and currently pass.  In the future, it may be
possible and more efficient to hold different formats of describing the
compilable layers, such as slices, but only index tuples are currently
used.

Note: compilable layer indices are not incorporated into circuit
conversion (to quil/qasm/cirq) yet.  This is left for future work.