Updates system tests to work with recent circuit updates.
Specifically the addition of compilable layer indices within a
Circuit object.  This commit regenerates the reference files used
by the system tests (since unpickling old circuits doesn't work
because a ._compilable_layer_indices_tup attribute is needed) and
makes minor updates to several of the tests.

The system tests pass on my laptop, so they should be close to passing
on GitHub.