Updates densitymx's computational state so tensor product bases are allowed.
A common way of creating explicit models creates a tensor-product basis
for multi-qubit models consisting of multiple 1Q 'pp' bases tensored
together.  This is also a fine Pauli-product basis, and should be accepted
as such within the basis-checking code in the computational state rep
in the densitymx and densitymx_slow evotypes.  This commit makes this
update, so tensor-product-of-'pp'-bases are allowed (in addition to a
pure 'pp' basis) when constructing a computational basis state.