Improves sub-data creation of CombinedExperimentDesign.
Filters the outcome labels of a combined design's data set when creating
a sub-data object when the outcome label is "simple" (no intermediate
outcomes) and contains just 0s and 1s.  In this case, the parent combined
design's dataset's outcome labels are filtered so that the sub-data object's
dataset only contains the outcome labels with the appropriate number of 0s
and 1s (e.g. if the sub-design's .qubit_labels is ('Q0','Q1') then only
length-2 outcome labels like ('00',) would be set as outcome labels of
the sub-data's dataset).

This fixes an issue that previously was dealt with by having the combined
design just create *all* the standard n-qubit outcome labels for the
appropriate number of qubits, but a recent commit reverted this because
it made subdata-creation take too long.  This commit should retain the
fast speeds but fix the issue whereby subdata datasets get polluted
outcome labels from the parent combined edesign, e.g. '0' and '1' as
well as '00', '01', etc.