First major WIP update of FOGI code.  Partial updates, nothing working yet.
Adds helper objecs such as an ElementaryErrorgenBasis, which ideally
will be helpful for holding the row/column bases used by the "gauge action"
matrices in the FOGI analysis.  Begins updates to the implementation,
but nothing is finished yet.  This commit gets to the point of creating
a common gauge space for all the operations, where we haven't decided
whether an intersect-local-gauge-spaces method will be prefereable to
just finding the nullspace of a potentially large sparse matrix like we
did with a dense matrix in the original implementation.