Updates local and cloud noise model creation to work when passed OpFactory members.
Previously whether a gate was considered a factory or not was based entirely on
whether the processor spec's 'gate_unitaries' element was callable.  Now the
gate unitary may be a numpy array but if the passed-in 'gatedict' element is derived
from OpFactory the model correctly places the object with its factories.

This commit also updates LabelTupWithArgs.strip_args() to return a LabelStr when
the original label's .sslbls is None, as this seems more appropriate.  This ended
up not being needed to get current code to work, but seems like a sensible behavior
and may be needed to handle factory label resolution in the future.