Updates FOGI analysis, mostly, to use sparse matrices and minimally work again.
ExplicitModel.setup_fogi works again, creating now sparse matrices of FOGI
directions using sparse math in some (but not all) areas.  In particular, the
lack of a good sparse nullspace routine means that dense matrices are used
after a point, hopefully at which the "gauge op" matrices and vectors are small
enough (not tested yet).  In any case, further upgrading to utilize sparse
matrices more should be straightforward and can be done incrementally.

One impedence mismatch is that the updated FOGI analysis uses (for better
scaling to more qubits) elementary error generators labeled via
(sslbls, ('H' or 'S', <basisEls without I>)) whereas the models and ops
still use the old ('H' or 'S', <basisEls with I>) format.  This is HACKed
currently in a routine in explicitmodel.py, that only works with Pauli
bases.  In the future we need to update all of the elementary errorgen
labels to use the FOGI type (and maybe make a class for these labels).

Still a WIP, but basic FOGI analysis functionality is working and seems
to match pre-update results.