Slight change to Basis.cast that avoid creating 1-element DirectSumBasis objects.
Now when casting a basis that would be just a tensor product (1 tensor-prod block),
e.g. Basis.cast('std', [(4,4)]), a TensorProdBasis is returned instead of a
DirectSumBasis that contains as its sole element a TensorProdBasis.  This seems
like a natural and useful behavior, but may break some tests that depend on the
result being a direct sum basis (?).