Adds initial SPAM support for FOGI quantities (explicit model only)
Largely untested, but adds state preparations and POVMs as "operations"
with associated gauge reference frames (`ccomms` in code) found by
ascertaining what gauge directions are "irrelevant" to a SPAM op
(are in it's null space, so A*rho=0 or Meas*A=0 and exp(A)*rho = rho, etc)
and taking the operators "gauge action" to be the identity on the
complement of this space - the "relevant" space.  SPAM operations are
assumed (hardcoded) to produce no intrinsic FOGI quantities and to
treat the entire "relevant" space as a gauge reference frame.  I don't
think these assumptions are always strictly true and expect we'll need
to tweak them in the future.  Still a WIP.