Adds and integrates ElementaryErrorgenLabel objects.
The new GlobalElementaryErrorgenLabel and LocalElementaryErrorgenLabel
objects encapsulate elementary errorgen labels with and without a set of
target state space labels, respectively.  "Local" labels are still used when
building error generator matrices, taking projections, and converting to/from
parameter arrays, but the global labels can be used to construct
LindbladErrorgen objects and will be returned by the .coefficients
and .errorgen_coefficients (and similar) methods of operation objects.

This removes ambiguity regarding the target state-space-labels and
non-uniformity that was present previously (e.g. coefficient labels from a
ErrorgenOp would be given a new format vs. a plain exponentiated
lindblad errorgen).  It also holds the support of the error generator
in a more useful way (e.g., rather than a string like "IIIIIIXIIIII"
it holds 'X' with a sslbls of (6,) ).

Usage is still largely untested, and may needs some further implementation.
FOGI routines, however, now use the same type of labels as the operations,
which will allow further completion of the FOGI framework.