Updates FOGI framework so that FOGI components can be *set* as well as retrieved.
Now that the same (global type of) elementary errorgen labels are used
for both the FOGI computations and in the operator objects, it is possible
to convert a given list of FOGI components to an errorgen-set space vector
and set a model's errors accordingly.  This commit enables this functionality,
and in the 1- and 2-qubit examples we've tried it successfully gets & sets
FOGI components correctly.

This commit also fixes a few minor errors (missing a .conjugate()) in
the SPAM FOGI analysis.