Fixes minor bugs caught by unit tests.
Mostly related to the switch to using GlobalElementaryErrorgenLabel
objects.  Also updates the basis construction test to expect that
a TensorProdBasis rather than a DirectSumBasis results when an Basis.cast
is specified with a single tensor product block (I don't think this
will break anything else, but we should be aware of it).