Works toward making more objects JSON serializable; updates metadir framework; new GSTGaugeOptSuite object.
Adds _to_memoized_dict and _from_memoized_dict for a number of classes that
are included in experiment designs and protocols (and more left to do, this is
a WIP!).  These methods implement object serialization to/from JSON or other
non-pickle files that should allow us to disconnect the Python object from the
serialized format, allowing us to upgrade/update the former without destroying
backward compatibility.

This commit also updates the meta-directory serialization interface, changing
the allowed "auxfile_types" to include "serialized-object" (which uses the new
to/from memoized dict methods) and improving the format & implementation of
serialzing container objects.  For example, the auxfile type for a dictionary
of protocol objects is now "dict:protocolobj" vs. the old "dict-of-protocolobjs".
The new implementation is more flexible and powerful.

Lastly, this commit merges the old "gaugopt_suite" and "gaugeopt_target" arguments
of several of the protocols (e.g. GST and ModelTest) into a GSTGaugeOptSuite object.
This simplifies the arguments given to these protocols, but more importantly allows
us to serialize this information nicely.