Adds paritial de-serialization for Models.
The ultimate goal is to make Model serializable in a way that doesn't
depend on their internals (e.g. when using pickle).  This commit adds
a number of from_memoized_dict methods to model and modelmember objects,
but serializing models will still require a that a number of other classes
be serializable - so this is a WIP.  Nothing tested yet.