Adds "memoized dict" serialization to multiple supporting objects.
Adds _to_memoized_dict and _from_memoized_dict to a number of object
types that are held by Model or Protocol classes.  This framework is
still a WIP, and hasn't even been tested yet.  This commit enables us
to start testing Model serialization (for Explicit and LocalNoise, not
CloudNoise models) and GST/ModelTest protocol serialization.