Adds on-demand ProtocolResultsDir child creation and several minor fixes.
Now a root ProtocolResultsDir can create sub-results on demand
instead of failing with a NotImplementedError when, for instance,
we load a root directory with a sub-directory that doesn't have results
but for which a sub-sub-directory does. Previously the child-loading
code would not load this sub-directory and it woudl be inaccessible.
Now it isn't loaded intially but will be loaded on demand if the parent
ProtocolResultsDir is indexed.

This commit also adds a number of minor bug fixes so that we can load
back a written edesign/data/results tree successfully - at least one
containing just GST results.