Adds serialization of sub-data cache items.  Fixes JSON bugs.
Allows data-cache items to be JSON'd instead of pickled, mainly by
adding to/from serialization methods to NamedDict.  During this update,
it was noticed that dicts with non-string keys are not actually JSON-able
(the keys are converted to strings!).  This commit adds a, perhaps
temporary, check for such insidious dicts before we write JSON and
fixes all the cases that it caught.