Updates POVMs fnd Instruments to use ModelMember.submembers framework.
This alters POVMs such that the constituent effect vectors' .gpindices
are now *global* parameter indices rather than indices local to the
POVM.  This matches the usage of "submembers" elsewhere and makes
the serialization of models more consistent and straightforward.
Similarly the members of instruments hold global (model-level) parameter
indices.

As a part of this, we implement a paradigm shift whereby a modelmember
now indicates that it is "unallocated" and needs to be allocated to
a model by having its .parent=None rather than its .gpindices=None.
This update allows the .gpindices of a modelmember to be initialized
even without a parent model (which is sometimes useful, e.g. for
keeping track of the numbrer of parameters).  The new
ModelMember.init_gpindices method initalizes the gpindices of an
object based on those gpindices present in the sub-members:
- if the sub-members all belong to the same  non-None parent model,
  then the object allocates itself to this model (and we check that no
  new parameters are needed)
- if all/some of the sub-members are unallocated (or allocated to
  a different model?) then the object is allocated to the 'None'
  model, which means there is no parent model yet, and sub-members
  should allocate their parameters independently of one another.
This framework should (hopefully) give lots of flexibility in
constructing models with interdependent members.

This commit simplifies and revamps the Model._rebuild_paramvec
function, which should now be more robust to adding & removing
model members.  In testing, it seemed that we may want to rebuild
a model at the end of the .set_all_parameterizations call, and
after .copy is called (?) to keep the paramvec up to date.

This commit also fixes a bug in the idle-tomography methods run
when generating a report, and updates the unit tests to better
catch idle tomography failures in the future.

Finally, this commit removes a number of .copy() methods from
model member classes, as these were unnecessary and actually less
flexible than the base class version which uses a memoized deepcopy
to perform the copy and reset the parent model to None.

Tests have been updated, and currently most/all ar passing (unit
and system).