Fixes a bug in densitymx evotype internals which could cause a segfault.
Updated the type of state used as a proxy in OpRep.aslinearoperator from
a base StateRep (which isn't actually initialized though it doesn't complain
about this because of special behavior of __cinit__) to a StateRepDense object.
This fixes a segfault error that would occur because the StateRep wouldn't
contain a pointer to valid data.

This commit also fixes this error everwhere else we could find this same
mistake, and additionally adds necessary arguments to StateRepDensePure
objects within the statevec evotype (and swaps the __init__ args of this
class to conform to the ordering elsewhere).