Updates I/O API function names to use "read" instead of "load".
For future proofing, I/O functions that work with filenames are now
"read" or "write" rather than "load" and "write", and the convention
for "write" functions is to have the first argument be the filename
and later arguments describe what is to be written.  Deprecated stub
functions have been added to direct users to the new functions and
signatures.