Updates reference model in order to fix a system test.
A recent commit updated the argument order of the
pygsti.evotypes.statevec.statereps.StateRepDensePure object, making
one of the reference model files used by the system tests unable to
be unpickled.  This commit simply regenerates that file so the
test_stdgst_terms test passes again.

This commit also updates the argument order to be correct in
for the statevec_slow evotype, which had been forgotten in the
earlier commit.