Refactors get_* functions and methods in pyGSTo to have other names.
As part of the pyGSTi code style we avoid "get_"-prefixed functions and
methods, as this often doens't add much meaning.  More importantly, we'd
like the function naming to be consistent.  This commit renames various
get_* functions, usually by simply removing the "get_" and sometimes
replacing with a "retrieve_" or other verb.