Updates MarginalizedPOVM so circuit.simulate just calls model.probabilities.
Previously, because of a bug in (or lack-of-features in) MarginalizedPOVM,
the Circuit.simulate function performed additional marginalization beyond
what was output by model.probabilities.  This was because MarginalizedPOVM
didn't distinguish between different orderings of the qubits being kept,
and so addition logic was needed to straigthen this out.  This commit updates
MarginalizedPOVM so that the ordering matters (as intuitively it should)
and removes the extra logic from Circuit.simulate so that apart from filtering
out near-zero probabilities it just calls model.probabilities.

Adds a unit test to ensure correct behavior in a simple 4-qubit case.