Fixes bug in lindblad_terms_to_projections so it can handle integers as basis labels
optools.lindblad_terms_to_projections takes a dictionary of Lindblad terms,
usually of the form ('H', 'XX'), but *also* allowed is the form ('H', 3) where
3 is taken to mean the [3] element of the also provided lindblad basis.  Such
integers weren't being properly converted to actual labels and downstream this
caused an equality check to fail, which is what caught this bug.  Fixed now.