Adds convenience behavior to unitary/pure ops/states.
Adds ability to construct a unitary op or pure state from a
super-op or super-ket, respectively.  This functionality isn't documented
yet and many not be needed.

Adds a fallback when creating Lindblad-type POVMs to create "static"
base effect vectors when "static unitary" fails.  This is a more
necessary and clearly useful update.