Enhances performance of lindblad_tools.nonham_lindbladian
This function gets called a lot when constructing LindbladErrorgen
operators (which get build during loading).  This commit replaces sparse
operations with dense ones when the size of the matrices being constructed
are small (udim <= 4, so 1-2 qubit), even when the `sparse` argument is set
to True, and the result is simply converted to a sparse matrix at the very
end.  This function could probably still be sped up further, and maybe even
cached efficiently with a little work.