Adds a "for deployment" version of pyproject.toml
The last attempt failed to build wheels, I think because when a
pyproject.toml file is used this makes the python build occur in a
different "place" (container/directory?) so that the work done in
CIBW_BEFORE_BUILD (installing cython and other python dependencies)
doesn't have effect on the python performing the build (the
wheel building setup.py run gave the warning that build tools weren't
installed).

This commit attempts to fix this by putting the (optional) build
dependencies like cython in an alternate pyproject.toml file
(named pyproject.toml.fordeploy) that is swapped in by the
CIBW_BEFORE_BUILD actions.  We'll see if this works...