Updates unit testing workflow, adds version=0.0 debug print statements.
Ok, now we have wheels being successfully built and uploaded to testpypi,
but with version number 0.0.  Statements added to testdeploy.yml should
help debug what's going wrong with setuptools_scm when we're building
wheels.

Secondly, since we updated to pyproject.toml the unit tests have
been failing (an errant __version__ somewhere, or maybe a package
needs __version__?) but more importantly cython extensions haven't
been being built.  I think this is because having the pyproject.toml
file causes a wheel to be built, and this is done in a separate
"build environment" where cython isn't installed unless it's in the
pyproject.toml as a build requirement.  This commit causes the unit
test script to use the deployment pyproject.toml (which includes
cython), and hopefully this will cause the unit tests action to
build cython extensions again.