Changes top-level attribute from 'version' back to '__version__', as this is more standard.
The _version.py file (created by setuptools_scm) defines a `version`, but this is now
imported as `__version__` by pygsti/__init__.py.  Using a dunder version attribute seems
more the standard, and we print this before running the unit tests, so it will fix this too.
I think this is needed by standard python packages.