Fixes hessian memory estimation and improves memory estimate reporting overall.
Fixes some incorrect "array_type" specifications that led to an overly
large (pessimistic) estimate of the memory required to evaluate the .hessian(...)
of an objective function ("local" quantities - meaning held on an entire processor
- were used when "block" quantities should have been, e.g. "epp" => "abb").

This commit also adds printing (logging) of the layout's estimated required
memory when using a matrix or map forwardsim.  It also leaves some commented-out
debugging code in matrixforwardsim.py for printing out the memory corresponding
to each array type that is required - this will be removed soon.

Also fixes atom counting in print statements when computing the hessian to
be 1-based instead of 0-based.