Adds caching to LindbladErrorgen._init_generators for performance.
Loading CPTP explicit models from disk appeared slow and bottlenecked
by _init_generators.  This commit adds customized caching of this
method based on the members of LindbladErrorgen that is utilizes,
storing the cache in a class member.  It also folds into _init_generators
the computation of generator 1-norms, as this was the next bottleneck.

As a related update, this commit fixes ExplicitBasis caching to actually
look at element values rather than the basis name, which if auto-generated
as ExplicitBasis_XXX never match each other.

Resulting speedup for a 2Q explicit model loading is 5-8x.  Further improvement
could be made if we used standard bases rather than explicit ones, as the
hashing is simpler (current bottleneck is checking basis equivalence).